many small updates - fix a category editing bug in EditGame - add a loading state to Library - remove deprecated code - extract string resources
