+++ app/src/main/java/com/waynebloom/scorekeeper/singleGame/statisticsForGame/StatisticsForGameConstants.kt
@@ -1,4 +1,4 @@
-package com.waynebloom.scorekeeper.singleGame.statisticsForGame.domain
+package com.waynebloom.scorekeeper.singleGame.statisticsForGame
 
 object StatisticsForGameConstants {
