--- app/src/main/res/values/edit_game_strings.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <string name="a11y_categories_close_button">Close scoring categories editor screen</string>
-    <string name="header_scoring_categories">Scoring Categories</string>
-    <string name="header_edit_categories">Edit Categories</string>
-    <string name="header_game_details">Game Details</string>
-    <string name="header_custom_theme">Custom Theme</string>
-    <string name="info_categories_section_helper">If your game has multiple scoring sources, add them here. An example set would be City, Farm, Road, and Monastery.</string>
-    <string name="scoring_option_ascending">Lowest Score Wins</string>
-    <string name="scoring_option_descending">Highest Score Wins</string>
-    <string name="scoring_option_manual">Decide Your Own Ranks</string>
-</resources>
