+++ app/src/main/res/values/strings.xml
@@ -68,14 +68,29 @@
 
     <!-- endregion -->
 
-    <!-- region GamesScreen -->
+    <!-- region LibraryScreen -->
 
     <string name="text_empty_game_search_results">No games found containing \"%s\"</string>
     <string name="text_no_game_name">No Name</string>
 
     <!-- endregion -->
 
-    <!-- region GameStatistics -->
+    <!-- region EditGameScreen -->
+
+    <string name="a11y_categories_close_button">Close scoring categories editor screen</string>
+    <string name="button_edit_categories">Edit Categories</string>
+    <string name="button_add_a_category">Add a Category</string>
+    <string name="header_scoring_categories">Scoring Categories</string>
+    <string name="header_game_details">Game Details</string>
+    <string name="header_custom_theme">Custom Theme</string>
+    <string name="info_categories_section_helper">Add scoring categories to your game. For example, in \'Settlers of Catan\' a player scores points for buildings, development cards, and the competitive objectives.</string>
+    <string name="scoring_option_ascending">Lowest Score Wins</string>
+    <string name="scoring_option_descending">Highest Score Wins</string>
+    <string name="scoring_option_manual">Decide Your Own Ranks</string>
+
+    <!-- endregion -->
+
+    <!-- region StatisticsForGameScreen -->
 
     <string name="header_scoring">Scoring</string>
     <string name="header_plays">Plays</string>
@@ -101,24 +116,20 @@
         <item quantity="other">%d wins</item>
     </plurals>
     <string name="number_list_overflow">…and %1$d more</string>
+    <string name="no_data_for_category">There is no data recorded for this category</string>
 
     <!-- endregion -->
 
-    <!-- region MatchesForSingleGameScreen -->
+    <!-- region MatchesForGameScreen -->
 
+    <string name="sort_by">Sort by</string>
+    <string name="sort_direction">Sort direction</string>
     <string name="sort_option_age">Date of Match</string>
     <string name="sort_option_name">Winning Player</string>
     <string name="sort_option_score">Winning Score</string>
     <string name="sort_option_players">Player Count</string>
     <string name="text_empty_match_players">No Players Added</string>
     <string name="text_empty_match_search_results">No matches found containing \"%s\".</string>
-
-    <!-- endregion -->
-
-    <!-- region Overview -->
-
-    <string name="header_recent_matches">Recent Matches</string>
-    <string name="top_bar_header_overview">Overview</string>
 
     <!-- endregion -->
